let env var overrule .dot env setting (#2136)
* let env var overrule .dot env setting

Summary: The script adlsgen2setup.py does not assign the "groups" ACL to individual files, and prepdocs.py does not check folder ACLs and created a copy of all folders in flat struture without any acls groups / rights. Consequently, the ACL is missing from the "group" field in the vector index —it remains empty. Files are accessible for anyone independent or the security group he is in

#2109

* Load azd env vars variable

---------

Co-authored-by: Pamela Fox <pamela.fox@gmail.com>